requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
blinker==1.4
cached-property==1.5.1
certifi==2018.10.15
chardet==3.0.4
Click==7.0
colander==1.5.1
coverage==4.5.1
decorator==4.3.0
docker==3.5.1
docker-compose==1.23.0
docker-pycreds==0.3.0
dockerflow==2018.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
filelock==3.0.9
flake8==3.6.0
Flask==1.0.2
gunicorn==19.9.0
idna==2.7
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.3.0
pluggy==0.8.0
py==1.7.0
pycodestyle==2.4.0
pyflakes==2.0.0
pytest==3.9.3
pytest-cov==2.6.0
pytest-flask==0.14.0
python-dateutil==2.7.5
python-decouple==3.1
pytz==2018.7
PyYAML==3.13
requests==2.20.0
six==1.11.0
texttable==0.9.1
toml==0.10.0
tox==3.5.3
translationstring==1.3
urllib3==1.24
virtualenv==16.0.0
websocket-client==0.53.0
Werkzeug==0.14.1
